Update microbenchmark scripts

+++ benchmarks/cluster-scripts/run_microbenchmark_job.sh
@@ -8,30 +8,37 @@ NONDETERMINISM=$5
 NUM_SHARDS=$6
 TASK_DURATION=$7
 MAX_FAILURES=$8
-OUTPUT_DIR=${9:-"$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"}
+COLLECT_LATENCY=$9
+OUTPUT_DIR=${10:-"$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"}
 
 HOME_DIR=$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )
 
 NUM_TASKS=100
 NUM_ITERATIONS=1
 
-if [[ $# -eq 8 || $# -eq 9 ]]
+if [[ $# -eq 9 || $# -eq 10 ]]
 then
     echo "Running with $NUM_RAYLETS workers, use gcs only? $USE_GCS_ONLY, $GCS_DELAY_MS ms GCS delay, nondeterminism? $NONDETERMINISM"
 else
-    echo "Usage: ./run_job.sh <num raylets> <head IP address> <use gcs only> <GCS delay ms> <nondeterminism> <num shards> <task duration> <num failures> <output dir>"
+    echo "Usage: ./run_job.sh <num raylets> <head IP address> <use gcs only> <GCS delay ms> <nondeterminism> <num shards> <task duration> <num failures> <collect latency?> <output dir>"
     exit
 fi
-output_prefix=$OUTPUT_DIR"/latency-"$NUM_RAYLETS"-workers-"$NUM_SHARDS"-shards-"$USE_GCS_ONLY"-gcs-"$GCS_DELAY_MS"-gcsdelay-"$NONDETERMINISM"-nondeterminism-"$TASK_DURATION"-task-"$MAX_FAILURES"-failures-"
+
+if [[ $COLLECT_LATENCY -eq 1 ]]
+then
+    output_prefix=$OUTPUT_DIR"/latency-"$NUM_RAYLETS"-workers-"$NUM_SHARDS"-shards-"$USE_GCS_ONLY"-gcs-"$GCS_DELAY_MS"-gcsdelay-"$NONDETERMINISM"-nondeterminism-"$TASK_DURATION"-task-"$MAX_FAILURES"-failures-"
+else
+    output_prefix=$OUTPUT_DIR"/lineage-"$NUM_RAYLETS"-workers-"$NUM_SHARDS"-shards-"$USE_GCS_ONLY"-gcs-"$GCS_DELAY_MS"-gcsdelay-"$NONDETERMINISM"-nondeterminism-"$TASK_DURATION"-task-"$MAX_FAILURES"-failures-"
+fi
 
 if ls $output_prefix* 1> /dev/null 2>&1
 then
-    echo "Latency file with prefix $output_prefix already found, skipping..."
+    echo "File with prefix $output_prefix already found, skipping..."
     exit
 fi
 
 output_prefix=$output_prefix`date +%y-%m-%d-%H-%M-%S`
-output_file=$output_prefix.txt
+output_file=$output_prefix.csv
 raylet_log_file=$output_prefix.out
 
 exit_code=1
@@ -59,21 +66,26 @@ do
       cmd=$cmd" --gcs-only"
     fi
 
-    echo $cmd | tee -a $output_file
-    $cmd 2>&1 | tee -a $output_file
+    echo $cmd
+    $cmd 2>&1
 
     exit_code=${PIPESTATUS[0]}
 
     i=$(( $i + 1 ))
 done
 
-
-echo "UNCOMMITTED LINEAGE" >> $output_file
-for worker in `cat ~/workers.txt`
-do
-  ssh -o "StrictHostKeyChecking no" -i ~/ray_bootstrap_key.pem $worker "grep 'UNCOMMITTED' /tmp/ray/*/logs/raylet.err" | awk -F':' '{ print $5 }' >> $raylet_log_file
-done
-
-sort $raylet_log_file | uniq -c | sort -bgr >> $output_file
-rm $raylet_log_file
-
+echo "Collecting stats from workers..."
+if [[ $COLLECT_LATENCY -eq 1 ]]
+then
+    echo "latency" >> $output_file
+    for worker in `tail -n $NUM_RAYLETS ~/workers.txt`; do
+        ssh -o StrictHostKeyChecking=no -i ~/ray_bootstrap_key.pem $worker "grep LATENCY /tmp/ray/*/logs/worker*" | awk -F'LATENCY:' '{ print $2 }' >> $output_file
+    done
+else
+    echo "uncommitted_lineage" >> $output_file
+    for worker in `tail -n $NUM_RAYLETS ~/workers.txt`; do
+      ssh -o "StrictHostKeyChecking no" -i ~/ray_bootstrap_key.pem $worker "grep 'UNCOMMITTED' /tmp/ray/*/logs/raylet.err" | awk -F':' '{ print $5 }' >> $raylet_log_file
+    done
+    sort $raylet_log_file | uniq -c | sort -bgr >> $output_file
+    rm $raylet_log_file
+fi
